# openapi.yaml
openapi: 3.0.1
info:
  description: Formación API Rest Meetup Deloitte DC
  title: API Rest Full
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://127.0.0.1:8080
    description: Servicio Local
...
paths:
...
  /v1/companies:
    get:
      operationId: getAllCompanies_2
      parameters:
        - in: query
          name: isMultiNational
          required: false
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniesV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompaniesV1'
          description: OK
      tags:
        - Company V1
    post:
      operationId: newCompany_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyV1'
          application/xml:
            schema:
              $ref: '#/components/schemas/CompanyV1'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV1'
          description: OK
      tags:
        - Company V1
  /v1/companies/{id}:
    delete:
      operationId: getOsiTipById_1
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: boolean
          description: OK
        '404':
          content:
            '*/*':
              schema:
                type: boolean
          description: Not Found
      tags:
        - Company V1
    get:
      operationId: getCompanyById_1
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV1'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV1'
          description: Not Found
      tags:
        - Company V1
    put:
      operationId: updateCompany_1
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyV1'
          application/xml:
            schema:
              $ref: '#/components/schemas/CompanyV1'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: boolean
            application/xml:
              schema:
                type: boolean
          description: OK
        '404':
          content:
            application/json:
              schema:
                type: boolean
            application/xml:
              schema:
                type: boolean
          description: Not Found
      tags:
        - Company V1
...
  /v1/countries/{id}:
    get:
      operationId: getCountryById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CountryV1'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryV1'
            application/xml:
              schema:
                $ref: '#/components/schemas/CountryV1'
          description: Not Found
      tags:
        - Country V1
  /v2/companies:
    get:
      operationId: getAllCompanies
      parameters:
        - in: query
          name: isMultiNational
          required: false
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniesV2'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompaniesV2'
          description: OK
      tags:
        - Company V2
    post:
      operationId: newCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyV2'
          application/xml:
            schema:
              $ref: '#/components/schemas/CompanyV2'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV2'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV2'
          description: OK
      tags:
        - Company V2
  /v2/companies/{id}:
    delete:
      operationId: getOsiTipById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: boolean
          description: OK
        '404':
          content:
            '*/*':
              schema:
                type: boolean
          description: Not Found
      tags:
        - Company V2
    get:
      operationId: getCompanyById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV2'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV2'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyV2'
            application/xml:
              schema:
                $ref: '#/components/schemas/CompanyV2'
          description: Not Found
      tags:
        - Company V2
    put:
      operationId: updateCompany
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyV2'
          application/xml:
            schema:
              $ref: '#/components/schemas/CompanyV2'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: boolean
            application/xml:
              schema:
                type: boolean
          description: OK
        '404':
          content:
            application/json:
              schema:
                type: boolean
            application/xml:
              schema:
                type: boolean
          description: Not Found
      tags:
        - Company V2
  /v2/companies/{id}/countries:
    get:
      operationId: getCountriesForCompany
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Countries'
            application/xml:
              schema:
                $ref: '#/components/schemas/Countries'
          description: OK
      tags:
        - Company V2
components:
  schemas:
    CollectionModelCompanyV2:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            companyV2List:
              type: array
              items:
                $ref: '#/components/schemas/CompanyV2'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelCountryV1:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            countryV1List:
              type: array
              items:
                $ref: '#/components/schemas/CountryV1'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CompaniesV1:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanyV1'
      xml:
        name: companies
    CompaniesV2:
      type: object
      properties:
        companies:
          $ref: '#/components/schemas/CollectionModelCompanyV2'
      xml:
        name: companies
    CompanyV1:
      type: object
      properties:
        id:
          type: string
        isMultinational:
          type: boolean
        name:
          type: string
        size:
          type: string
      xml:
        name: company
    CompanyV2:
      type: object
      properties:
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        id:
          type: string
        isMultinational:
          type: boolean
        name:
          type: string
        size:
          type: string
      xml:
        name: company
    Countries:
      type: object
      properties:
        countries:
          $ref: '#/components/schemas/CollectionModelCountryV1'
      xml:
        name: countries
    CountryV1:
      type: object
      properties:
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        id:
          type: string
        name:
          type: string
      xml:
        name: country
    Link:
      type: object
      properties:
        deprecation:
          type: string
        href:
          type: string
        hreflang:
          type: string
        name:
          type: string
        profile:
          type: string
        templated:
          type: boolean
        title:
          type: string
        type:
          type: string